Deduplicate MemberInfoTests
